Refactored combat plugin layout
